+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"/Users/meredithdavies/northcoders/project/nc-final-project-frontend/.eslintrc.js":"1"},{"size":292,"mtime":1647859815759,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","suppressedMessages":"6","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"fze6t7","/Users/meredithdavies/northcoders/project/nc-final-project-frontend/.eslintrc.js",[],[]]
