+++ .eslintignore
@@ -1,1 +1,1 @@
-/__tests__/*
+/__tests__/*
