+++ components/Footer.jsx
@@ -1,0 +1,1 @@
+import React, { Component } from "React";
