+++ package.json
@@ -8,11 +8,13 @@
     "ios": "expo start --ios",
     "web": "expo start --web",
     "eject": "expo eject",
-    "prepare": "husky install"
+    "prepare": "husky install",
+    "test": "TZ=UTC jest"
   },
   "dependencies": {
     "expo": "~44.0.0",
     "expo-status-bar": "~1.2.0",
+    "firebase": "^9.6.9",
     "jest": "^27.5.1",
     "react": "17.0.1",
     "react-dom": "17.0.1",
